change int to decimal but not sure everything okey

+++ StockProductTracking/MVVM/Model/Product.cs
@@ -13,8 +13,8 @@ namespace StockProductTracking.MVVM.Model
         public int CategoryId { get; set; }
         public string ProductTitle { get; set; }
         public string ProductBrand { get; set; }
-        public int ProductPrice { get; set; } //satış
-        public int ProductRealPrice { get; set; } //alış
+        public decimal ProductPrice { get; set; } //satış
+        public decimal ProductRealPrice { get; set; } //alış
         public int ProductStock { get; set; }
